De-leak run_inbox message mapper from k8s + tighten scenario 04 (warren-5af5)
- Move toSeamMessage(RunInboxRow -> Message) out of src/runtime/k8s/
  send-message.ts into the neutral src/runs/inbox-message.ts, so the
  local/shared poll path (src/runs/inbox.ts) no longer imports from the
  k8s tree (same pattern as warren-9574's authenticatedCloneUrl).
- Scenario 04: assert burrow.workspacePath is exactly "" instead of the
  weaker any-string check, catching a regression that would leave a
  stale host path or empty an unexpected field.
